#2238 fixed memory leaks
