+++ app/src/main/java/org/kiwix/kiwixmobile/di/components/KiwixActivityComponent.kt
@@ -35,7 +35,6 @@ import org.kiwix.kiwixmobile.settings.KiwixSettingsActivity
 import org.kiwix.kiwixmobile.splash.KiwixSplashActivity
 import org.kiwix.kiwixmobile.webserver.ZimHostActivity
 import org.kiwix.kiwixmobile.webserver.ZimHostModule
-import org.kiwix.kiwixmobile.zim_manager.fileselect_view.ZimFileSelectFragment
 import org.kiwix.kiwixmobile.zim_manager.fileselect_view.effects.DeleteFiles
 import org.kiwix.kiwixmobile.zim_manager.library_view.LibraryFragment
 
@@ -49,7 +48,6 @@ import org.kiwix.kiwixmobile.zim_manager.library_view.LibraryFragment
 )
 interface KiwixActivityComponent : CoreActivityComponent {
   fun inject(libraryFragment: LibraryFragment)
-  fun inject(zimFileSelectFragment: ZimFileSelectFragment)
   fun inject(readerFragment: KiwixReaderFragment)
   fun inject(localLibraryFragment: LocalLibraryFragment)
   fun inject(deleteFiles: DeleteFiles)
