+++ app/src/main/java/org/kiwix/kiwixmobile/nav/destination/library/LocalLibraryFragment.kt
@@ -18,6 +18,8 @@
 
 package org.kiwix.kiwixmobile.nav.destination.library
 
+import android.Manifest
+import android.content.pm.PackageManager
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.Menu
@@ -25,30 +27,149 @@ import android.view.MenuInflater
 import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.view.ActionMode
 import androidx.appcompat.widget.Toolbar
+import androidx.core.content.ContextCompat
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import io.reactivex.disposables.CompositeDisposable
+import kotlinx.android.synthetic.main.fragment_destination_library.file_management_no_files
 import kotlinx.android.synthetic.main.fragment_destination_library.go_to_downloads_button_no_files
+import kotlinx.android.synthetic.main.fragment_destination_library.zim_swiperefresh
+import kotlinx.android.synthetic.main.fragment_destination_library.zimfilelist
 import org.kiwix.kiwixmobile.R
 import org.kiwix.kiwixmobile.cachedComponent
 import org.kiwix.kiwixmobile.core.base.BaseActivity
+import org.kiwix.kiwixmobile.core.base.BaseFragment
 import org.kiwix.kiwixmobile.core.extensions.ActivityExtensions.start
+import org.kiwix.kiwixmobile.core.extensions.ActivityExtensions.viewModel
+import org.kiwix.kiwixmobile.core.extensions.toast
 import org.kiwix.kiwixmobile.core.main.CoreMainActivity
 import org.kiwix.kiwixmobile.core.utils.LanguageUtils
+import org.kiwix.kiwixmobile.core.utils.REQUEST_STORAGE_PERMISSION
+import org.kiwix.kiwixmobile.core.utils.SharedPreferenceUtil
 import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.adapter.BookOnDiskDelegate
+import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.adapter.BooksOnDiskAdapter
+import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.adapter.BooksOnDiskListItem
 import org.kiwix.kiwixmobile.local_file_transfer.LocalFileTransferActivity
+import org.kiwix.kiwixmobile.zim_manager.ZimManageViewModel
 import org.kiwix.kiwixmobile.zim_manager.ZimManageViewModel.FileSelectActions
 import org.kiwix.kiwixmobile.zim_manager.ZimManageViewModel.FileSelectActions.RequestMultiSelection
 import org.kiwix.kiwixmobile.zim_manager.ZimManageViewModel.FileSelectActions.RequestNavigateTo
 import org.kiwix.kiwixmobile.zim_manager.ZimManageViewModel.FileSelectActions.RequestSelect
 import org.kiwix.kiwixmobile.zim_manager.fileselect_view.FileSelectListState
-import org.kiwix.kiwixmobile.zim_manager.fileselect_view.ZimFileSelectFragment
+import javax.inject.Inject
 
-class LocalLibraryFragment : ZimFileSelectFragment() {
+private const val WAS_IN_ACTION_MODE = "WAS_IN_ACTION_MODE"
 
-  override val bookDelegate: BookOnDiskDelegate.BookDelegate by lazy {
+class LocalLibraryFragment : BaseFragment() {
+
+  @Inject lateinit var viewModelFactory: ViewModelProvider.Factory
+  @Inject lateinit var sharedPreferenceUtil: SharedPreferenceUtil
+
+  private var actionMode: ActionMode? = null
+  private val disposable = CompositeDisposable()
+
+  private val zimManageViewModel by lazy {
+    requireActivity().viewModel<ZimManageViewModel>(viewModelFactory)
+  }
+
+  private val bookDelegate: BookOnDiskDelegate.BookDelegate by lazy {
     BookOnDiskDelegate.BookDelegate(sharedPreferenceUtil,
       { offerAction(RequestNavigateTo(it)) },
       { offerAction(RequestMultiSelection(it)) },
       { offerAction(RequestSelect(it)) })
+  }
+  private val booksOnDiskAdapter: BooksOnDiskAdapter by lazy {
+    BooksOnDiskAdapter(bookDelegate, BookOnDiskDelegate.LanguageDelegate)
+  }
+
+  override fun onDestroyView() {
+    super.onDestroyView()
+    actionMode = null
+    disposable.clear()
+  }
+
+  override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+    super.onViewCreated(view, savedInstanceState)
+    zim_swiperefresh.setOnRefreshListener(::requestFileSystemCheck)
+    zimfilelist.run {
+      adapter = booksOnDiskAdapter
+      layoutManager = LinearLayoutManager(context, RecyclerView.VERTICAL, false)
+      setHasFixedSize(true)
+    }
+    zimManageViewModel.fileSelectListStates.observe(viewLifecycleOwner, Observer(::render))
+    disposable.add(sideEffects())
+    zimManageViewModel.deviceListIsRefreshing.observe(viewLifecycleOwner, Observer {
+      zim_swiperefresh.isRefreshing = it!!
+    })
+    if (savedInstanceState != null && savedInstanceState.getBoolean(WAS_IN_ACTION_MODE)) {
+      zimManageViewModel.fileSelectActions.offer(FileSelectActions.RestartActionMode)
+    }
+
+    disposable.add(zimManageViewModel.libraryTabIsVisible.subscribe { finishActionMode() })
+
+    go_to_downloads_button_no_files.setOnClickListener {
+      offerAction(FileSelectActions.UserClickedDownloadBooksButton)
+    }
+  }
+
+  private fun finishActionMode() {
+    actionMode?.finish()
+  }
+
+  private fun sideEffects() = zimManageViewModel.sideEffects.subscribe(
+    {
+      val effectResult = it.invokeWith(requireActivity() as AppCompatActivity)
+      if (effectResult is ActionMode) {
+        actionMode = effectResult
+      }
+    }, Throwable::printStackTrace
+  )
+
+  private fun render(state: FileSelectListState) {
+    val items: List<BooksOnDiskListItem> = state.bookOnDiskListItems
+    bookDelegate.selectionMode = state.selectionMode
+    booksOnDiskAdapter.items = items
+    if (items.none(BooksOnDiskListItem::isSelected)) {
+      actionMode?.finish()
+    }
+    actionMode?.title = String.format("%d", state.selectedBooks.size)
+    file_management_no_files.visibility = if (items.isEmpty()) View.VISIBLE else View.GONE
+    go_to_downloads_button_no_files.visibility = if (items.isEmpty()) View.VISIBLE else View.GONE
+  }
+
+  override fun onResume() {
+    super.onResume()
+    checkPermissions()
+  }
+
+  override fun onSaveInstanceState(outState: Bundle) {
+    super.onSaveInstanceState(outState)
+    outState.putBoolean(WAS_IN_ACTION_MODE, actionMode != null)
+  }
+
+  private fun checkPermissions() {
+    if (ContextCompat.checkSelfPermission(
+        requireActivity(),
+        Manifest.permission.READ_EXTERNAL_STORAGE
+      ) != PackageManager.PERMISSION_GRANTED
+    ) {
+      context.toast(R.string.request_storage)
+      requestPermissions(
+        arrayOf(Manifest.permission.READ_EXTERNAL_STORAGE),
+        REQUEST_STORAGE_PERMISSION
+      )
+    } else {
+      requestFileSystemCheck()
+    }
+  }
+
+  private fun requestFileSystemCheck() {
+    zimManageViewModel.requestFileSystemCheck.onNext(Unit)
   }
 
   private fun offerAction(action: FileSelectActions) {
@@ -93,18 +214,5 @@ class LocalLibraryFragment : ZimFileSelectFragment() {
     activity.setupDrawerToggle(toolbar)
     setHasOptionsMenu(true)
     return root
-  }
-
-  override fun render(state: FileSelectListState) {
-    super.render(state)
-    val items = state.bookOnDiskListItems
-    go_to_downloads_button_no_files.visibility = if (items.isEmpty()) View.VISIBLE else View.GONE
-  }
-
-  override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-    super.onViewCreated(view, savedInstanceState)
-    go_to_downloads_button_no_files.setOnClickListener {
-      offerAction(FileSelectActions.UserClickedDownloadBooksButton)
-    }
   }
 }
